Export ExtendedAdapterInfo properly
